+++ CHANGELOG.md
@@ -19,7 +19,7 @@ All notable changes to this project will be documented in this file.
 * Remove User non-essential display methods ([#852](https://gitlab.com/yale-sdmp/vesta/issues/852)).
 * Change DrawPolicy#group_actions to remove policy check against draft stage of the draw ([#722](https://gitlab.com/yale-sdmp/vesta/issues/722)).
 * Hide the Terms of Service link when user is not logged in ([#978](https://gitlab.com/yale-sdmp/vesta/issues/978)).
-
+* Allow export of all group members' emails (no longer restricted to group leaders only) ([888](https://gitlab.com/yale-sdmp/vesta/issues/888))
 
 ### Fixed
 * Fix AdminDashboard membership creation failure ([#971](https://gitlab.com/yale-sdmp/vesta/merge_requests/1018)).